Remove dependance on CMS.
This site had some content coming from an external CMS
that was pulled in at build time. This commit removes dependance on that CMS and
instead (at least temporarily) stores this data statically in the code. This
will allow modification of this content that was previously in a CMS with edits
of the file (src/static-cms.ts) in git and a PR.

* Removed `getStaticProps` as it is not needed anymore.
* Removed a number of unused util files.
* Pulled in the data that was previously fetched via a CMS.
    * Put a cut down and cleaned up version of this in `src/static-cms.ts`
    * Put a raw unmodified version of this in `cms-data-dump.jsonc` for
      reference.
* Pulled in images that were pulled from the CMS and put them in the
  `public/home-page` directory
* There was a system for handling CMS previews that was removed.
* There was a system for threading news banner info through the site that was
  removed as we reference it directly now.
* Removed axios as we are no longer making network calls.
* Added types for `react-lottie` as they were missing.
* Removed all the `pages/api` handlers as they are now unused.

Note: This is a first pass to allow for iteration now that the CMS is gone.
      Things can be improved from here.

Signed-off-by: Michael Goff <thephred@gmail.com>